Update the item controller function to handle GET/POST requests for the item create form
This commit creates an item form view that renders on a GET request for the form. During the POST request, the form values are validated and sanitized. If there are errors, the create form is re-rendered with the clean values that displays the error messages. If there are no errors, a new item is created and saved into the data or, if the item already exists, the user is redirected to the item detail page.