Update the category controller function to handle GET/POST requests for the category delete form
This commit creates a category delete form view that renders on a GET request for the form. If a category contains items, a list of the items are displayed to instruct the user to delete the items before deleting the category. If the category does not contain items, submitting a POST request to delete the category will successfully delete the category from the database.